Checkers should derive from CheckerBase, typehint this in Interpreter
