+++ src/Script/Interpreter/BitcoinCashChecker.php
@@ -3,15 +3,22 @@
 namespace BitWasp\Bitcoin\Script\Interpreter;
 
 use BitWasp\Bitcoin\Script\ScriptInterface;
-use BitWasp\Bitcoin\Serializer\Transaction\TransactionSerializer;
 use BitWasp\Bitcoin\Transaction\SignatureHash\Hasher;
 use BitWasp\Bitcoin\Transaction\SignatureHash\SigHash;
 use BitWasp\Bitcoin\Transaction\SignatureHash\V1Hasher;
 use BitWasp\Buffertools\Buffer;
 use BitWasp\Buffertools\BufferInterface;
 
-class BitcoinCashChecker extends Checker
+class BitcoinCashChecker extends CheckerBase
 {
+    /**
+     * @var array
+     */
+    protected $sigHashCache = [];
+
+    /**
+     * @var int
+     */
     protected $sigHashOptionalBits = SigHash::ANYONECANPAY | SigHash::BITCOINCASH;
 
     /**
@@ -26,22 +33,18 @@ class BitcoinCashChecker extends Checker
             throw new \RuntimeException("SigVersion must be 0");
         }
 
-        $cacheCheck = $sigVersion . $sigHashType . $script->getBuffer()->getBinary();
+        $cacheCheck = $sigHashType . $script->getBuffer()->getBinary();
         if (!isset($this->sigHashCache[$cacheCheck])) {
             if ($sigHashType & SigHash::BITCOINCASH) {
                 $hasher = new V1Hasher($this->transaction, $this->amount);
             } else {
-                if ($this->hasherV0) {
-                    $hasher = $this->hasherV0;
-                } else {
-                    $hasher = $this->hasherV0 = new Hasher($this->transaction, new TransactionSerializer());
-                }
+                $hasher = new Hasher($this->transaction);
             }
 
             $hash = $hasher->calculate($script, $this->nInput, $sigHashType);
             $this->sigHashCache[$cacheCheck] = $hash->getBinary();
         } else {
-            $hash = new Buffer($this->sigHashCache[$cacheCheck], 32, $this->adapter->getMath());
+            $hash = new Buffer($this->sigHashCache[$cacheCheck], 32);
         }
 
         return $hash;
